Add Test For Insufficient RAM State
The new failure condition in the `start` hook should be properly tested to ensure it works properly, so this commit adds such a test. The test itself revealed that the failure wasn't being detected properly to begin with, so it was good to have, as it allowed the check to be fixed. To simplify writing the test, the `start_container` function in `util/docker.sh` was extended to accept an optional third argument for the memory size, which defaults to the 256 which was in use for all containers before the value was bumped up to let Travis have all tests pass. To prevent this from being a testing regression, the two tests which had been failing before that commit were both updated to pass the larger size in the new argument. Finally, the `configure` hook was updated to place upper and lower bounds on the memory setting, both to prevent ES from failing to start (the lower bound), and to maintain performance levels across varying environments (the upper bound). A comment was added explaining that change, as well as the selection process used for the values of the bounds.